playwright-python/requirements.txt
playwright
pytest-playwright
pytest
pytest-html
pytest-xdist
PyYAML